+++ src/main/java/org/octri/common/validation/CurrentOrPastDate.java
@@ -1,0 +1,27 @@
+package org.octri.common.validation;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
+
+/**
+ * Annotation that adds {@link CurrentOrPastDateValidator} validation to an entity field.
+ */
+@Documented
+@Constraint(validatedBy = CurrentOrPastDateValidator.class)
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.FIELD)
+public @interface CurrentOrPastDate {
+
+	String message() default "Date may not be in the future";
+
+	Class<?>[] groups() default {};
+
+	Class<? extends Payload>[] payload() default {};
+
+}